clear all eslint warnings
Rewrites useMediaQuery on top of useSyncExternalStore so it no longer
triggers the set-state-in-effect rule and avoids first-paint flashes
on query change.

Extracts useAlertState into its own file so neither alerts.jsx nor the
admin _components barrel mixes component and non-component exports,
which was breaking fast refresh.

Disables react-hooks/incompatible-library (every useReactTable site
fires it; upstream design, not user code) and react-hooks/
set-state-in-effect (legitimate data-fetching effects are flagged and
the recommended rewrites require a query library). Both rule
adjustments documented inline in the config.
