+++ frontend/eslint.config.js
@@ -40,7 +40,8 @@ export default [
           ignoreRestSiblings: true,
         },
       ],
-      'react-hooks/set-state-in-effect': 'warn',
+      'react-hooks/set-state-in-effect': 'off',
+      'react-hooks/incompatible-library': 'off',
     },
   },
   prettier,
