+++ frontend/src/component/alerts.jsx
@@ -1,4 +1,3 @@
-import { useCallback, useRef, useState } from 'react';
 import { Alert } from '@optiaxiom/react';
 
 export const AlertBanner = ({ alert, onDismiss }) => {
@@ -8,28 +7,4 @@ export const AlertBanner = ({ alert, onDismiss }) => {
       {alert.message}
     </Alert>
   );
-};
-
-export const useAlertState = (autoDismissMs = 4000) => {
-  const [alert, setAlert] = useState(null);
-  const timerRef = useRef(null);
-
-  const dismiss = useCallback(() => {
-    if (timerRef.current) clearTimeout(timerRef.current);
-    timerRef.current = null;
-    setAlert(null);
-  }, []);
-
-  const show = useCallback(
-    (intent, message) => {
-      if (timerRef.current) clearTimeout(timerRef.current);
-      setAlert({ intent, message });
-      if (autoDismissMs && intent !== 'danger') {
-        timerRef.current = setTimeout(() => setAlert(null), autoDismissMs);
-      }
-    },
-    [autoDismissMs]
-  );
-
-  return { alert, show, dismiss };
 };
