+++ frontend/src/component/useMediaQuery.js
@@ -1,17 +1,22 @@
-import { useEffect, useState } from 'react';
+import { useCallback, useSyncExternalStore } from 'react';
+
+const getServerSnapshot = () => false;
 
 export const useMediaQuery = (query) => {
-  const [matches, setMatches] = useState(() =>
-    typeof window === 'undefined' ? false : window.matchMedia(query).matches
+  const subscribe = useCallback(
+    (callback) => {
+      if (typeof window === 'undefined') return () => undefined;
+      const mql = window.matchMedia(query);
+      mql.addEventListener('change', callback);
+      return () => mql.removeEventListener('change', callback);
+    },
+    [query]
   );
 
-  useEffect(() => {
-    const mql = window.matchMedia(query);
-    const handler = (event) => setMatches(event.matches);
-    setMatches(mql.matches);
-    mql.addEventListener('change', handler);
-    return () => mql.removeEventListener('change', handler);
-  }, [query]);
+  const getSnapshot = useCallback(
+    () => (typeof window === 'undefined' ? false : window.matchMedia(query).matches),
+    [query]
+  );
 
-  return matches;
+  return useSyncExternalStore(subscribe, getSnapshot, getServerSnapshot);
 };
